feat: create a delete modal component for inventory-process
Based on Martijns component

Refs: oph-716
